* changes:
  Rootcanal: do not starve the reader in unittest
  Factor out duplicated warning options
  RootCanal: Report AES CCM encryption if both devices support Secure Connections
  RootCanal: return the correct mode in Read Scan Enable
  RootCanal: Implement filtering of Inquiry packets based on IAC
  RootCanal: Add basic LMP encryption tests
  [Rootcanal] Add PAGE_TIMEOUT for classic connections on Rootcanal
  RootCanal: Bypass LL CONNECTION PARAMETER REQ in central role
  RootCanal: Implement additional filter accept list parameter validation
  floss: Add rerun-if-changed cargo instruction when using bt_packetgen
  Fix Rootcanal extended inquiry response generation
  Add some LMP SSP backward compatibility procedures test cases
  Add some LMP Secure Simple Pairing Passkey Entry Notification tests
  Add some LMP Secure Simple Pairing failure tests
  RootCanal: Implement additional resolving list parameter validation
  RootCanal: Fix parsing of arrayed controller properties
  Add some LMP Secure Simple Pairing OOB tests
  Add some LMP Secure Simple Pairing Passkey tests
  RootCanal: legacy advertising PDU updates
  Fix setting of std::optional variables.
  Update source for Rust 1.63.0
  RootCanal: Implement ECDH for BR/EDR
  RootCanal: Properly determine link key type
  RootCanal: Move feature loader to features.rs
  RootCanal: Move link key notification to pairing
  Update for Rust v1.62.0

+++ system/gd/dumpsys/bundler/test.cc
@@ -66,7 +66,7 @@ TEST_F(BundlerTest, CreateBinarySchemaBundle) {
   std::vector<flatbuffers::Offset<bluetooth::dumpsys::BundledSchemaMap>> vector_map;
   std::list<std::string> bundled_names;
   ASSERT_TRUE(CreateBinarySchemaBundle(&builder, filenames, &vector_map, &bundled_names));
-  ASSERT_EQ(0, vector_map.size());
+  ASSERT_EQ((unsigned int)0, vector_map.size());
 }
 
 TEST_F(BundlerTest, WriteHeaderFile) {
